% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setMethods_torch.R
\name{c.GPUmatrix}
\alias{c.GPUmatrix}
\title{Concatenate GPUmatrix objects}
\usage{
\method{c}{GPUmatrix}(...)
}
\arguments{
\item{...}{objects to be concatenated}
}
\description{
Concatenate GPUmatrix objects
}
