% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{multiplyLowerDouble}
\alias{multiplyLowerDouble}
\title{Multiply lower triangular matrix}
\usage{
multiplyLowerDouble(C, A, B, Nglobal0, Nlocal0, NlocalCache)
}
\arguments{
\item{C}{output matrix}

\item{A}{lower triangular matrix}

\item{B}{rectangular matrix}

\item{Nglobal0}{number of global work items}

\item{Nlocal0}{number of local work items}

\item{NlocalCache}{rows of B to cache in local memory}
}
\description{
Multiplies a lower triangular matrix by a rectangular matrix
}
