% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatisticalApplications.R
\name{summary.GPUglm}
\alias{summary.GPUglm}
\title{Summary for GPUglm objects}
\usage{
\method{summary}{GPUglm}(object, correlation = FALSE, ...)
}
\arguments{
\item{object}{A GPUglm object}

\item{correlation}{Logical, if TRUE the correlation matrix of the estimated parameters is returned and printed. Default is FALSE.}

\item{...}{Additional arguments}
}
\description{
Summary for GPUglm objects
}
