\name{xtfrm}
\title{Auxiliary Function for Sorting and Ranking}
\docType{methods}



\alias{xtfrm-methods}
\alias{xtfrm,gpu.matrix.tensorflow-method}
\alias{xtfrm,gpu.matrix.torch-method}


\description{
Mimics \code{xtfrm} basic function.
}


\usage{

\S4method{xtfrm}{gpu.matrix.tensorflow}(x)
\S4method{xtfrm}{gpu.matrix.torch}(x)

}


\arguments{

\item{x}{A GPUmatrix object}
}


\seealso{
\code{\link[base]{xtfrm}}
}

